Improve file vs folder saving GUI for better UX
When text is changed in a metadata field, two buttons now appear next to the field allowing the user to save the change either to the file or to the entire folder. At the bottom of the pane, a 'Save' button remains in place, which saves all of the text contents of all of the metadata fields to the file.